polished the comparefq script a bit

+++ bin/comparefq.sh
@@ -12,7 +12,7 @@ die() {
 
 if [ $# -ne 2 ]
 then
-	echo "Usage: ${0} guppy_basecall_dir1 guppy_basecall_dir2"
+	echo "Usage: ${0} a.fastq b.fastq"
 	exit
 fi
 
@@ -22,7 +22,11 @@ B=$2
 test -e $A || die "$A not present."
 test -e $B || die "$B not present."
 
+test -e a.fastq && die "temporary file a.fastq is present. Remove that first!"
+test -e b.fastq && die "temporary file b.fastq is present. Remove that first!"
+
 cat $A  | paste - - - -  | sort -k1,1 -T . | tr '\t' '\n' > a.fastq
 cat $B  | paste - - - -  | sort -k1,1 -T . | tr '\t' '\n' > b.fastq
 diff -q a.fastq b.fastq || die "fastq differ"
 
+rm -f a.fastq b.fastq
